I just scrapped and restarted
And now I'm making even more progress. How is that even possible? Maybe I know more about what needs to be done.

Overall, still not ready for servers. I'm in the middle of working on encryption. I stopped here with the error I got with connecting to the client. Will look into that tomorrow see what needs done but I'm far too tired right now to fix it. Just wanted to show my progress of where I'm at.

Packet receiving/sending has been massively improved. It's so much more easier and a lot less messy, I'll say.

The only thing so far that is unique to this software is the basis of the player database. The Player Object/DB will be more in depth in times to come but for now this is what it is while I get the protocol sorted out.

I've also kept the configuration at bay as long as I can, I won't add it in until I absolutely need to. Hopefully I can get through the protocol part pretty quickly because after that the main features of the software will be a breeze to develop including the commands.