Update Checking & Config
A system has been put in place to check for updates.

Also put in place was the config, which will now be updated as we move on with the project. I decided this feature was important to add now rather than later because of how expansive and huge the protocol is. It would have been a pain to go in and modify everything already done if I did the config later.