All Play Packets into the enum
All the play packets (serverbound and clientbound) have been entered into the Packet enum. This way they will be ready when I go to call on them individually. It will be important to note that the serverbound ones that have the same name as a clientbound one have attached at the end "ServerBound" to their name.