Just An Update, A Lot of Errors
A lot of errors here, just wanted to update this for people to see where I am at in the project.

This part is not tested, there are a lot of errors in the code because what I'm working on is NOT finished. Essentially I am working on porting GemsCraft-Classic over to this new software.